+++ docs/changelog.rst
@@ -29,6 +29,7 @@ v0.10.0 (in development)
 
 - Warn on encountering a repository version with a greater minor version than
   expected
+- Gave `PyPISimple` a `~PyPISimple.download_package()` method
 
 
 v0.9.0 (2021-08-26)